mysite/requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4
typing_extensions==4.9.0

pip==24.2
djangorestframework>=3.12
Pillow>=8.0
numpy>=1.19
pandas==2.2.2